add tree mapper into transaction namespace
