+++ src/transaction/transaction.ts
@@ -61,6 +61,7 @@ export class Transaction {
 
     instances.forEach((ins, idx) => {
       this.trackProperties(ins);
+      this.trackPredicates(ins);
 
       const predicates = MetadataStorage.Instance.predicates.get(ins.constructor.name);
       if (!predicates) {
@@ -70,7 +71,6 @@ export class Transaction {
       // FIXME: If the same uid is referenced in multiple places in the data, currently we will have 2 different instances
       //   of the same object. We need to make sure we share the instance.
       predicates.forEach(pred => {
-        this.trackPredicate(ins, pred);
         const _preds = (plain[idx] as any)[pred.args.name];
 
         if (_preds) {
@@ -102,6 +102,15 @@ export class Transaction {
     });
 
     return;
+  }
+
+  private trackPredicates<T extends Object, V>(instance: T): void {
+    const predicates = MetadataStorage.Instance.predicates.get(instance.constructor.name);
+    if (!predicates) {
+      return;
+    }
+
+    predicates.forEach(p => this.trackPredicate(instance, p));
   }
 
   private trackPredicate<T extends Object, V>(instance: T, metadata: PredicateMetadata): void {
@@ -164,5 +173,98 @@ export namespace Transaction {
     getSetNQuads: (target: T) => ISetMutation<Quad[]>;
     getSetNQuadsString: (target: T) => ISetMutation<String>;
     getDeleteNQuads: () => string;
+  }
+
+  export function of<T>(entryType: Constructor<T>): TreeMapperBuilder<T> {
+    return new TreeMapperBuilder().addEntryType(entryType)
+  }
+
+  class TreeMapperBuilder<T = any> {
+    private _entryType: Constructor<T>;
+    private _jsonData: IObjectLiteral<any>[];
+    private _resource = new Map<string, IObjectLiteral<any>>();
+
+    addEntryType(type: Constructor<T>): TreeMapperBuilder<T> {
+      this._entryType = type;
+      return this;
+    }
+
+    addJsonData(data: IObjectLiteral<any> | IObjectLiteral<any>[]): TreeMapperBuilder<T> {
+      this._jsonData = Array.isArray(data) ? data : [data];
+      return this;
+    }
+
+    /**
+     * Walk the resource graph and add all nodes into resource cache by its `uid`.
+     */
+    addResourceData(data: IObjectLiteral<any> | IObjectLiteral<any>[]): TreeMapperBuilder<T> {
+      if (data && !(data instanceof Array) && data.uid) {
+        this._resource.set(data.uid, data);
+        return this;
+      }
+
+      data.forEach((d: any) => {
+        this.addResourceData(d);
+      });
+
+      return this;
+    }
+
+    build(): Transaction.IEnvelope<T> {
+      const context = new Transaction();
+      // Do not traverse the json tree if there is no
+      // resource data.
+      if (this._resource.size > 0) {
+        const visited = new Set<string>();
+        Array.isArray(this._jsonData)
+          ? this._jsonData.map(jd => Private.expand(visited, this._resource, jd))
+          : Private.expand(visited, this._resource, this._jsonData);
+      }
+
+      if (!Array.isArray(this._jsonData)) {
+        this._jsonData = [this._jsonData];
+      }
+
+      const instances = context.transform(this._entryType, this._jsonData);
+      instances.tree.forEach(i => context.diffTracker.purgeInstance(i));
+      return instances;
+    }
+  }
+}
+
+/**
+ * Module private statics
+ */
+namespace Private {
+  /**
+   * Visit all nodes in a tree recursively, matching node uid in the resource data and adding extra information.
+   *
+   * ### NOTE
+   * Expand will modify the data in-place.
+   */
+  export function expand(visited: Set<string>, resource: IObjectLiteral<any>, source: IObjectLiteral<any>): void {
+    if (resource.has(source.uid)) {
+      Object.assign(source, resource.get(source.uid));
+    }
+
+    Object.keys(source).forEach(key => {
+      if (key === 'dgraph.type') {
+        return;
+      }
+
+      if (!Array.isArray(source[key])) {
+        return;
+      }
+
+      source[key].forEach((node: any) => {
+        const visitingKey = `${source.uid}:${key}:${node.uid}`;
+        if (visited.has(visitingKey)) {
+          return;
+        }
+
+        visited.add(visitingKey);
+        return expand(visited, resource, node);
+      });
+    });
   }
 }
